Bug 581414 - Fix marking occurrences at end of source
Fix marking occurrences at the end of the source by excluding the source
itself from highlighting.

When mark occurrecnes is enabled, Xtext tries to mark all references to
the selected element, including the name of the element itself. When at
the end of the source, however, the selection resolves to the top-level
STSource. It then checks if the element has a name, which some of the
subclasses of STSource do have (to facilitate scoping). It then tries to
highlight the significant portion of the element (i.e., the name). Since
the name is only for the purpose of scoping, it does not exist in the
corresponding gammar. As a result, the whole element is marked, which in
this case happens to be the whole source. By telling Xtext that STSource
can never be referenced locally, this change stops it from considering
STSource to begin with.

Change-Id: I97464b86ca731552d26b0bb5b411d37f00bb108f
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>