Bug 581413 - NatTable's Type Selection Dialog refactoring
Some functionalities like the content provider have been extracted into
separate classes. The button for the Selection dialog has been added to
the tables a bit differently to avoid dependency issues. Project search
was taken from the search plug-in. The hierarchy of the directories
displays correctly (similarly to what we had for jface selection
dialog).

Change-Id: I8dfc49b0e341940119de1c9a47dae19cf27d74ed