Bug 581493 - removed doubled container moveing on content moving
When expanded subapp or group content was moved the
ResizeGroupOrSubappCommand was checking if a resize was needed. In
addition the original resizing code in the containercontent layoutpolicy
was also performing this action. This could lead to a double move and
with that a wrong position of the group.

This fix removes the doubled resize checks. All resize commands are now
handled in the ContainerContentLayoutPolicy. This reduces some
duplicated code and reduces complexity in the
FBNetworkXYLayoutEditPolicy.

Change-Id: I8ff55f7fd2e8a94bb210ab406060d1b796cb768c