Bug 581507 - Wrong movement of fbs of groups in expanded subapps
FBs contained in groups which themselves are contained in expanded
subapps where moved when the outer expanded subapps was resized on the
left and/or top. The reason is that groups have no containment.

The fix filters the FBs so that only the position of the correct FBs is
adjustes.

Change-Id: Id3da5e7d1a7c7f6dd0b1bc8d7b9daeab190dba09