bug 581513 - Resize expanded subapps on pin addition
When pins are added to expanded subapps checks are performed to
accordingly resize the subapp when needed. For that a new command is
introduced.

In order to make this cleaner and also better handleable the
CreationCommand which was just a marker interface was turned into an
abstract class. Which should now be a bit cleaner.

Change-Id: Ibaefc7185c45bc2c4d28c9eddbe913bab362a08e