Bug 579702 - remove data type library
In the first days of the interpreter, the blocks were loaded as an XMI.
As a result, the data types were null and had to be fetched from a data
type library. Nowadays, we use the 4diac system and type importers. As a
result, we should never have null-data types.

Change-Id: Ia507cf6c09f0a4504825e182a3a1b5e2c3648a5b