Bug 580270 - Reworked adapterfb handling for subapps
Had to introduce a second attribute to adapter decls for referencing to
adapterfbs which are shown in the fbnetwork of CFBs. The reason is that
the fbnetwork as well as the adapterfb property of adapter declarations
had the containment flag set. Moving the adapter fb to the network set
that adapterfb null. With a new attribute, adapterNetworkFB, is now used
for this case.

Change-Id: I6ebabd9bbeea6e88459c24a08a4e73521ab44686