Bug 581658 - New plugin, import wizard and connection to GitLab
Couple of things have been added: a new plugin by the name of
org.eclipse.fordiac.ide.gitlab. The classes it has (for now) are those
needed for the wizard and a download manager. The connection to GitLab
should work with a correct URL and an appropriate personal access token
(permissions set to api). However, for now I'm using a hard-coded
project ID (371) which will be changed in the upcoming commits when we
refine the wizard. Download of the files will follow. Project specific
settings have been removed.

Change-Id: Ibef2b27eddd3f7b72f69e980365ce36d5452c108
Signed-off-by: Dunja Zivotin <dunja.zivotin@jku.at>