Bug 581313 - ranking of emf matcher is wrong
The ranking is responsible for deciding which matcher "wins the race"
for a certain matching strategy. Unfortunately, the Factory sets all
rankings to 20. I believe that this is wrong, because in the plugin.xml
the ranking is defined higher - but that is completely ignored. The
platform calls the method setRanking(whatever_was_defined_in_xml) and
the value is just overridden to 20 hard-coded.

Change-Id: I2f790126a9b2fff60c69f86ad02ea40396fcef3b