Bug 581678 - connection auto layout executes only one command
Previously, the connection auto layout mechanism executed the connection
layout command after the triggering command itself. Since both commands
were executed on the command stack, the editor registered two changes.
This meant that the user had to undo and redo two times for each
performed action. Now the layout command is kept in the editor and not
executed on the command stack, which leads to the usability one would
expect.

Change-Id: I93f1d3d58768441bffe60bb37671b8baea3c42d3