Bug 581635 - avoid overriding type value with empty value
Parameters and values from connections need to be used for the
calculation. However, when these values are empty, the type initial
value is used for the computation. If even the type initial value is
absent, then it's up to the initial value of the data type.

Unfortunately, the type var's value was overridden with the value from
the network even when this value was null or empty - as a result, the
type initial value was lost. This fix ensures that we check the value
before writing it into the fb type.

Change-Id: I40872615dd2c8869bbb09d7aac6d1722c074f8ff