Bug 581675 - Fix duplicate document in ST algorithm refactoring
The refactoring framework uses a FileEditorInput to request the document
from the document provider. The prior implementation did not consider
FBTypeEditorInput and FileEditorInput equal, despite both referring to
the same file. This would lead the document provider to load a separate
document, which would then cause a desync between the original document
in the editor and the second document used by the refactoring.

This commit adapts the hashCode and equals methods of FBTypeEditorInput
to yield the same result as FileEditorInput and consider both equal if
referring to the same file. This allows the document provider to
correctly find and use the original document.

Change-Id: I23436c004926f9f98f8be25af5a602513c3ffe0d
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>