Bug 581761 - Remove and undo needs to preserve position in list
added index saving for fbnetworkelement and connection delete command.
When removing an elements/connection.. save and undo than the element
always gets placed at the end of the list containing all their types.
This can lead to different xml for the same project and also cause merge
conflicts. Therefore now when deleting the index in the list gets stored
and will be placed there when undoing.

Change-Id: I1eb2e0baddaba722f74402103f5be270703de6ab
Signed-off-by: fabio.gandolfi <fabio.gandolfi@jku.at>