Bug 581824 - Introduced SingleRecursiveContentAdapter
In order to get rid of the bad behavior of EContentAdapter this
singleREcursiveContentAdapter will attach itself only to the first level
of childrens of an object. This is for example needed for the pin
property sheet to show updates when direct editing a pin value.

For property sheets that need more recursion (e.g., interface tables)
dedicated but limited EContentAdapters are now used in addition.

During checking all property sheets most of them where cleaned for Java
17 instanceof pattern matching.

Change-Id: I27594e7a14d371ebee1176cb4c9ab2614538c941