Bug 581859 - Scan type files to detect if type name update is needed
Especially when importing or opening projects the resource change
listener can not distinguished if types are copied or otherwise added.
As loading the type in this cases may be to early as the typelib is not
yet correctly set up the resource change listener shouldn't load the
type for checking if the typename in XML is equal to the file name.

This fix scans the type file for the type name and directly checks the
typename without fully parsing it. Only if the name is different the
file is parsed and updated.

Change-Id: I11be0f53604a39796f60a03ed163aa41d63d377d