Bug 582006 - Changed the Update Fbtype instances method to use the updateFB Type
In order to update the instance of an fbType correctly we can use the
UpdateFBType command.
Only untyped subapps need special treatment.
Furthermore, the update methods are moved into the same display thread.
Otherwise we will get troubles. For example the corresponding type is
not updated before the instances as we cannot know when async methods
are exucuted

Change-Id: Id53553ebcfe6ed43475b377d55408e732fb66b68