Bug 582233 - Extend tooltips for labels of hidden connections
Additional information was added in the tooltip for labels of hidden
connection: destination block + destination pin, comment of the
destination pin and a group (if the destination block is
a part of one). Added the same info for input hidden connections. Fixed
trailing whitespaces and allowed the same for single hidden connections.

Change-Id: Ieb92a9907d9e6d112a881b679bfa49b44caee114
Signed-off-by: Dunja Zivotin <dunja.zivotin@jku.at>