Bug 582244 - lock group size
Added the possibility to lock a group's size in it's property sheet. The
locked attribute determines if the auto-resize can change the group
width and height. Manual changes to the group size, e.g. dragging or
entering the size in the property sheet, are still allowed no matter the
value of the locked attribute.

Change-Id: If85074914afd1c0fb2e4cce149c0dbb9ab41182b