Bug 582325 - Reduce visibilty of monitoring context menu entries
Monitoring context menu entries where shown in different places when
there was a valid active editor. Even when the user was clicking in a
seperate view.

With this change now the visibility checks for monitoring context menus
are sharpend to be only shown when the user clicks in valid fb networks.

Change-Id: I159c5a83eba856a25014021cc8908b172d8a441e