Bug 582355 - Refactor assignment grammar in ST editor
The ST grammar used to distinguish between assignment and call
statements, requiring the use of a syntactic prediacte, since the
initial expression is identical. This led to several problems when
encountering syntax errors and when providing completion proposals.

This change unifies both assignments and ordinary expressions, including
calls, as statements. It also makes it possible to use multiple
assignments as a single statement.

Change-Id: I428cc7a9af968e5fb84300e28e39c5de50236065
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>