Bug 582378 - Add own font values for Diagram Font
Zooming in text editors changed the text font. This font was the basis
for the diagram font. As such zooming in text editors changed the font
size in the graphical editors leading to wrong graphics.

This fix adds own font values for the Diagram Font to separate graphical
and textual editors. The font values are based on the current setting in
the Eclipse platform.

Change-Id: I47db67c33ad02c40d00cdac1e32b1a35a658be53