Bug 582402 - Fix redundant calculations for initial values in nat table
The CalculatedValueCache from nattable does not check, whether there is
currently a calculating executing for a specific key. As a result, it
queues a large number of redundant calculations when accessed before the
first respective calculation finishes.

This change adds a SafeCalculatedValueCache that properly guards agains
redundant calculations by using a set of currently executing keys.

Change-Id: Ia40f24556930e9325dff58016a19ce2b4d6fc560
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>