Bug 582347 - Added User choice to delete Refactoring preview
With this change the user now has the possibility select an adequate
update operation during the delete Structured type refactoring
operation. When a FBType contains a struct, the user now can choose
whether the Pin should be deleted or it should be changed to any struct.
After changing the types it has to be propagated to the instances as
well.

Change-Id: Ie2156de9659827c9e363e57055dd8072c2cfeaa7