[fix] Do not apply editable disabled style to headers of nat table
The current implementation uses the same editable rule for decorating
the editable disabled style of contents and headers. This causes a
mismatch between the column and row position in the config label
accumulator and the expected indices in the editable rule.

This change does no longer apply the editable disabled style for headers
altogether. If such styling is desired in the future, a separate
editable rule for the header should be used.

Change-Id: I6a92ca1897643028cf11a7778266f4aca338f18a
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>