[perf] optimize ST editor type synchronization
Avoid unnecessary type updates due to inconsequential changes in
the FB type or changes caused by the ST editor reconciler itself.

Also increase default delay for FB type update in ST editor to make
it different from (and also not a direct multiple) of the reconciler
delay. This avoids a potential resonance effect in the bidirectional
synchronization.

Change-Id: I8060f8f262bb9ac6aac3935324ff7b28d5eb36ec
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>