Bug 582489 - Open Project creates wrong errormarkers
changed the projectAdd workspace job to not clear the whole typelibrary
and fill it again but only add the newly added types via addition
finding. This fixed the problem causing missing datatypes, because the
workspace job is often slower than the system importer and it can be
that the typelib at that moment is not fully filled.

Change-Id: I67c83f5aaaa6bbb0db07e1caf340f25971df2086
Signed-off-by: fabio.gandolfi <fabio.gandolfi@jku.at>