Bug 582594 - Add EmptyWorkspaceHelper to SystemExplorer
The EmtpyWorkspaceHelper is the Eclipse Platform Infrastructure for
showing action links and welcome text in CommonNavigator-based Views
when there are no entries. With this we can now show new 4diac IDE
project and import wizard links.

Attention: EmptyWorkspaceHelper is an Eclipse Platform internal class.
This may imply API changes. However as the impact of any change on this
class is limited to our SystemExplorer class the impact would be low.

Change-Id: Ie9bfc02181d7c3b12be56938641f5c9e0789cd17