Bug 582610 - Moved StructEditingComposite to StructuredType
Instead of the data type entry the StructuredType is now used as the
main data class in the editing composite for structs. This makes it
better reusable for other tasks.

Furthermore some clean-up was performed that makes the UI generation
more self-contained and better follows SWT practices. This also
simplifies its usage in other contexts (e.g.,
AttributeDeclarationEditor).

Change-Id: I2bdb6e71ab8a76192710276d6423f8cf41f36427