Bug 582640 - ensure that group members are not copied on group copy
When using marquee select group members could be part of the selection
and resulting in duplicated paste entries.

A similar issue happend in cut and paste where entries where after
pasting not inside of their subapp or group.

We have now a doubled duplicate check. This is that future users of
paste are protected.

Change-Id: Ia7ec3885a9a3846ac2aab9b2d71a1ef55be4bba9