Bug 582670 - AdapterFB correctly added for CFBs
When creating the AdapterFB for a new plug or socket in a CFB the
AdapterFB was correctly added to the FBNetwork but the connection
between plug/socket and the AdapterFB got lost. This was for example
visible when renaming a plug/socket and the inner name was not correctly
updated.

Change-Id: If946b9fcf417f6a1eafa7024509fba731db62fe8