Bug 582675 - Fix trap for call statements after grammar reorganization
The ST core grammar has been reorganized to treat ST expressions as
statements. This included the removal of the separate ST call statement,
which results in the debugger no longer stopping on calls used as
single statements.

This commit fixes that by instead trapping on all expressions used as
statements (i.e., assignments and calls). As a result, most traces
in the evaluator tests needed to be changed, as well.

Change-Id: If6001dc484ed659dcb905c226a6fcdfdaf7a8da0
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>