Bug 532356 - Fix XML parsing with full app name in mapping
FB Names in resources with full app hierarchy name are now correctly
parsed, connections are resolved.

Furthermore the mapping parsing has been reworked to allow both the
extended mapping syntax (previously used by 4diac IDE) and the reduced
mapping syntax (to has only dev and res).

With this change from now has to be an application path and to a
resource path as it is defined in the standard.

Change-Id: I6799e2ee590a016ee18ea5f5add4a7e344c6e26d