Bug 582729 - Remove Validate Project from SystemManager
Currently we perform a a full project and type load for the whole
workspace on 4diac IDE startup. This comes from the time we did model
validation during loading the xml files. With the new builder based
approach this is not necessary and useful anymore and costs only startup
performance.

Change-Id: Ibf4dd0f372b9337bf26806874fab16f3f108505b