Bug 582837 - Open expanded subapps in their next editable parent
As expanded subapps are internally opened in a viewer the double click
(or single click with for link with editor) on an expanded subapp in the
system explorer opens the next editable parent and selects the subapp.

As part of this change also the double click of FBs leads to the right
place and the goto menu handles expanded subapps the same way.

Change-Id: Iedca8641830fd6f955bd2b8c19014eb1d2b1ca3b