Bug 582854 - Fix attribute declaration embedded editor context
The embedded editor used the closest container of type LibraryElement as
the context, such as for retrieving packages and imports. For an attribute
declaration, this meant that the intermediate DataType, which is also a
LibraryElement, was used instead of the outermost AttributeDeclaration
containing the actual package and imports. This change now uses the root
container instead.

Change-Id: I4af837b68c75f068057b01eb5d62f4a6ecfafa00
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>