Bug 582892 - Improve Drawing Performance of Connections in FBNs
The different paint methods of Draw2d figures are pushing and saving
graphics state and setting clip areas. This can produce a lot of
garbage. Especially when the connection is hidden.

Therefore this commit is reducing the paint operations of
FBNetworkConnections as much as possible. Also no state is saved. For
FBNetworkConnections this is fine as we are always drawing to the
Connection layer and the layer's paint method is performing the required
state preservation.

Change-Id: I9067598805d78700d1fdec5e197c4a761f1bb77c