Bug 582900 - Handle file changes synchronously by ResourceChangeListener
For large changes in the files system Eclipse may invoke the
ResourceChangeListener several times. Therefore all actions because of
one change invocation have to be completed before the next one.
Especially handing stuff of to the display thread can not be done
asynchronous if data of the ResourceChangeListener is used.

Change-Id: I194518f07d1ba11fb20203ea356d302b99b289d6