Bug 582931 - Fix deadlock when refactoring data types
The UpdatePinInTypeDeclarationCommand did a separate save of the updated
type, which lead to a deadlock between the UI thread performing the save
in the current editor holding a lock on the type entry and the ongoing
refactoring operation holding a lock on the workspace.

Also performed some clean-up on the command itself.

Change-Id: Ida21cff6ae0ed56cdb56402993dbf3a13bec810e
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>