Bug 582945 - Save sub-editors only if dirty in FB type editor
There was a misaligned text replace on rename refactoring of an
interface element. The change to the interface element saves the FB type
editor, which unconditionally saved all child editors. The ST algorithm
editor then triggered an auto-format of its contents on saving, which
resulted in misaligned positions for the following ST change.

This change saves sub-editors only if they are dirty, avoiding the
auto-format and subsequent misaligned text replacement.

Change-Id: I2a65c536ecfddd232e1ff2b294a4cad946838ba6
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>