Bug 583013 - REmvoed UnfoldedSuabppContentNetwork
With the extended use of the EMF validator the use of an proxy object
like the UnfoldedSuabppContentNetwork made problems. Furthermore the
main reason for introducing such an object was the group. However for
subapps it is not necessary. With this commit directly the subapp's
fbnetwork is used. Simplyfying code in several places.

Change-Id: I5892624de4a90b83528eddb3640bec3aeb7a8d56