fix usage of List.getLast in Xtend code for Java 21
Java 21 adds List.getLast(), which throws an exception if the list is
empty. Xtend also has an extension method getLast(), which returned
null. Since the Java 21 method is declared directly in List, it has
precedence over the extension method, which changes the behavior of
existing Xtend code. This change fixes the problematic cases where
getLast is called in Xtend code.