Bug 583095 - Pin to pin connection for expanded subapps
In a pin to pin connection creation setting for expanded subapps the
connection was always added to the outside of the expanded subapp. As
the expanded subapp can now not be edited when going inside it was not
possible to create cut through connections (i.e. connections inside of
the subapp going from an input to an output pin of the expanded subapp).

With this change depending if the connection is created from input to
output or output to input, the connection is either created inside or
outside.

Change-Id: Ibb13774feac5d613326eb62720f214e0aa48ba3b