[f_move] fix saving f_move attribute
Although the specific methods for configurable fb's were included in the
model, they were not correctly generated and replaced with default code
in a recent patch. This patch therefore extracts the code from the
model, redefines the methods to be more generic, and moves the
StructManipulation class to the .impl-package because the methods should
not be used directly.

Change-Id: I0000000000000000000000000000000000000000