[demux] add visible children list to the model
Previously, the StructuredType field was used to store a copy of a
struct. This has proven to cause problems. Hence, a list of
visiblechildren will be added directly to the model. This should
facilitate handling configured demuxes. To distinguish a demux that was
not configured from one that does not have any children at the moment,
we need an extra flag to mark it as configured. This was previously
achieved by asking for the "VisibleChildren" attribute.

Change-Id: I2ed76bd055e15b6e070deee9b0e9e967f751fa76