[demux] add error marker for missing struct
In some cases (e.g., file deleted in the background via Git), the data
type of a structured type does not exist, although it is configured like
that in the mux/demux. Errors are now handled more gracefully and shown
in the graphical UI as red flags, similar to interface error pins.

Change-Id: I733471907cca55d2cc5da49a4802a1a95849e5f5