[demux] fix for loading visible children attribute
Previously, the output list of a demux was not properly cleared. This
change ensures that only the visible children are listed. Also correctly
adds the parent names now to the vardeclaration. Test had to be fixed
because attributes are not anymore stored all the time, checking the
attributes as they would be exported helps ensure that saving works
correctly.

Change-Id: I3250c584d4bc28d6108ee07e93517ae7208d71a5