[demux] fix parameter list in struct members
Because the name was not saved in the .-notation anymore, the parameters
were not correctly stored and interface elements were not reliably found
anymore. This patch overrides the getName() for membervardecls to avoid
too many special cases. In return, the new getVarName() only returns the
name of the element itself. Also fixes transferring parameters for
configurable FBs.

Change-Id: I7dbb6005612a1d87e56e71ced0452c6e9a2327ba