[demux] dismiss visible children that don't exist anymore
When a variable is deleted from a struct, that is configured in a demux,
we have to update the demux. In case the demux is preconfigured, it will
actually try to create an interface element based on the configuration
in the visible children attribute. Getting rid of these children is
currently only possible in the XML. It makes more sense to dismiss them
from the configured demux (otherwise each configured demux needs to be
adjusted, leading to a high knock-on viscosity)

Change-Id: I63d22a46b14f82c273653e81fc60cd5bb2905481