Fix convert subapp to group canExecute
The canExecute method is composed of a flatten subapp command and a
create group command (compoundcommand). At the time of the canExecute
check, the flatten subapp command was not yet executed, meaning that the
check for the netowrk of the subapp elements could not work in the
create group command.