fix ctrl+click for ECC to create state plus transition
When pressing ctrl while dragging an ECC transition, a new state is
created. This allows to quickly draw ECCs. With the recent change on the
position, this caused now an NPE. Additionally, the state is now created
directly under mouse position again (it had an offset for some time
already).