Any to Any connections crossing interfaces shall not be flagged as error
This change now will not mark any to any connections that are either
leaving a type (e.g., going to the interface of a CFB or typed subapp)
or that are crossing a untyped subapp, as an error. This is needed as
you may want to encapsulate a generic SIFB with some logic but have
generic data to be used (e.g., Client with response timeout checking).