Avoid concurrent mod. exception for configurable fb property sheets
Muxer and F_MOVE property sheets require a command stack listener to
correctly keep the selection and values when the data type is changed.
However in the listening method changes to the command stack listener
list where performed. With this change this is moved to an asynchronous
execution avoiding that.