Interface pins where not correctly added in expanded subapp viewers
when going into an expanded subapp we are now showing a non editable
viewer of the expanded subapp content. In these viewers the interface
pin of the expanded subapp where shown as it was an expanded subapp
inside of the subapp. this lead to connection artifacts from the outside
to be shown inside.