Extend open type in editor to struct and adapter pins
With this change open type in editor is also correctly shown for struct
and adapter pins. For pins of event type and elementary data type pins
the context menu entry is grey.

The menu entry in the Navigate menu works the same.

Addresses: https://github.com/eclipse-4diac/4diac-ide/issues/55