Use IEC 61499 Coord. System for Conn Bendpoints in Model
In order to minimize noise when saving on different screens we now store
the bendpoints of connections in IEC 61499 coordinate system in the
model. Only when connections are reposition the bendpoints are adjusted.
Furthermore we only store two digits in the xml file. This should reduce
noise in the change history.