Use IEC 61499 Coord. System for Element sizes
With this change we use the IEC 61499 coordinate system units also
inside of the model. For drawing the coordinates are translated to
screen units and back. This removes rounding errors. Together with
storing now the values with two digits after the comma it should reduce
noise in the XML files.