Improve Loading of TypeEditable
Loading the typeeditable sofar copied the type model. However copying
takes for larger types app the same amount of time then loading from
xml. So if the type was not loaded, loading type editable costed twice
the time.

This fix now only copies if the type is loaded and in the other case
loads it from the file.