do not save visible children attribute unnecessarily
With recent changes, there was a problem that the visible children
attribute was saved even when the demux was not configured. The current
change tries to fix this problem by checking first whether the attribute
is the default attribute - in which case the attribute is not exported.
Additionally, the visible children attribute is not stored for new FBs
anymore (check for isIsConfigured). Hence, also the string comparison is
only done for configured FBs