Fixed paste position for drag and drop paste in subapps
When dragging and dropping with ctrl inside of subapps the position of
the dropped element was wrong. This fix reduces unnecessary complexity
and is also correct when doing drag and drop and copy paste inside of
expanded subapps and groups.

What is not yet working is copy drag and drop into subapps and groups.
There still a position offset is happening.