Correctly create array for struct members
When creating the interface pins of a mux/demux, the array size was not
correctly initialized. This was added in the code for creating a
membervardeclaration from a given struct member vardeclaration. Note
that an array size is nowadays a model object with containment, so a
simple setArraySize(arraySizeObj) will break the containment of the
model. Hence, ECoreUtil.copy is used.