Refactoring of Safe Struct Deletion
Some changes had been built up in a wrong way.
This produced bugs, for example for having a struct in an interface 2
changes have been produced, one for deleting and one for updating.
This was contradicting.

The structure of the changes is reworked and refactoring
This improves usability for the user and ensures correctness.