Ensure that Property Sheet Refresh Does not Result in Changes #293
Refreshing a property sheet my result in the triggering of modify
listeners of our widgets. We ensure that according commands are not
executed by caching the command stack. This had to be done by each
property sheet individually, and if forgotten led to update issues
(e.g., segment property sheet).

With this fix this is handled in a common infrastructure in
AbstractSection.

https://github.com/eclipse-4diac/4diac-ide/issues/293