Fix ambiguous type resolution in attribute name column
The AttributeColumnAccessor did attempt to deresolve the new attribute
name with respect to the imports of the current type. This could lead
to the wrong attribute declaration being used when there were two
attribute declaration with the same name (but different package) in
the current scope. An example would be an attribute declaration in the
current package and one without any package, both with the same name.

This changes the name handling so that the entered name is used as-is.
The determination whether to use a full or unqualified name and whether
to add an import is now performed when creating the proposals.