Zooming via Toolbar Combobox did a Different Scrolling
When zooming in a graphical editor we either try to keep the position
under the mouse, or if the mouse is outside of the editor the center of
the current view, constant.

When zooming via the combobox this was different and always scrolled to
the center of the whole canvas. This is now harmonized.