ZoomFit considers Selection
When now objects are selected the zoom all toolbar item will zoom to the
bounds of the selection and position the viewport such that the top left
is the top left of the selection bounds.

Furthermore a bug in positioning the viewport when zoom all is invoked
without selection is also fixed.