show subapps instead of selecting them
when opened in the system explorer, expanded subapps were treated like
FBs, i.e,. they were selected and shown in the editor.

This is likely not the behavior one expects because most use cases do
not involve selecting an expanded subapp but rather working with its
content, such as shifting a selected group of FBs into an expanded
subapp. Hence, this commit reworks the behavior to actually show the
top-left corner and not changing the selection.