Escape URL when converting to URI (LibraryImportTest)
In #418 a workaround was introduced to escape space-characters during
the URL to URI conversion.
As https://docs.oracle.com/javase%2Ftutorial%2F/networking/urls/creatingUrls.html
points out: this should be done using the full 7-parameter constructor
of URI by passing the parts of the URL/URI to be constructed; this will
also handle other special characters and makes sure that encoding only
happens in the correct parts of the URI (e.g. `:` is not encoded where
it is used as a port-seperator, `@` is not encoded when it seperates the
user-info part,...)