Fix race condition with built-in data types in ST editor
The Xtext framework installs an adapter when getting qualified names of
elements via the qualified name provider. This causes a race condition
with the list of adapters for built-in data types in the ST editor.

This commit addresses the problem by directly providing the (qualified)
name without going through the qualified name provider for global data
types.