Deprecate type editable in type entries
The "editable" type may not be identical to the type currently used in
editors due to automatic reload from disk on changes. However, it was
often used this way.

This deprecates the type editable in favor of either getting a private
copy via TypeEntry.copyType() or getting the type directly from an
editor via Adapters.adapt(editor, LibraryElement.class).