New Launcher that allows to Stop 4diac FORTEs
The new launcher requires an 4diac FORTE executable and will stop all
currently local running 4diac FORTEs, which where launched from 4diac
IDE and have the given executable.

Furthermore to reduce code duplication an AbstractLaunchProcess class
was extracted from EvaluatorProcess and used for this launch as well as
for the DeploymentProcess.