Reconnect Connection Preserves Connection in XML File
In order to reduce the amount of changed lines in the XML file this
change ensures that reconnect connection will use the same places as the
old one. In the best case this means only source or target entry and
bendpoints are changed and such the changes are within one XML file
line. This should reduce chances for conflicts when multiple people work
on one file via version control.