Remove check for connected subapp pin in change data type command
The check for disallowing to change the type of a connected subapp pin
in the change data type command is frequently causing problems for
refactorings or quickfixes. The option to skip the check is often
forgotten.

This change removes the check entirely, while the checks in the UI that
prevent the user from changing the type still remain in place.