ECCEditorEditDomain required the Model to early #672
With the recent changes in managing the models a type editor is working
on the model is not available at the time the editdomain is created.
However this is also not really needed. This fix now moves the request
to the ECC model to the point when it is really needed.

Bug: https://github.com/eclipse-4diac/4diac-ide/issues/672