Remove Comment Value Escaping
Our XML writer is correctly escaping attribute values. If we are also
escaping them we get double escaped attribute values which are then
wrong. The escaping method was originally introduced for other use
cases. These are not valid anymore. A simple escaping of new lines and
removing of \r is sufficient.

Furthermore the comment attribute is only written if there is a non
blank comment. Reducing the resulting XML file size.