Fix tree-locked exception when creating markers after update in typelib
There is an exception when creating markers in the type library after a
resource update from the resource change listener, because the tree is
still locked for modifications. This commit makes sure that we are not
only holding the required lock but also that the tree is not locked when
performing marker updates synchronously.