Fix stale or ambiguous method parameter references in ST editor
The current resource in the ST editor also contains the ST algorithms
and methods of the copied internal FB type and their parameters. These
were subsequently included in the object descriptions of the resource,
which caused stale or ambiguous references compared to the actual ST
algorithms and methods in the ST source. This filters the ST algorithms
and methods of the internal FB type, including their children, when
computing the descriptions to avoid the problem.